Fix for non-empty folder accidental delete. (#239)
* Add check to delete_fcd, check if folder is empty
And only delete if the folder is empty after the fcd_delete vmware task is finished

+++ cinder/volume/drivers/vmware/volumeops.py
@@ -2240,7 +2240,10 @@ class VMwareVolumeOps(object):
                                         id=fcd_location.id(cf),
                                         datastore=fcd_location.ds_ref())
         self._session.wait_for_task(task)
-        if delete_folder:
+        folder_path = f"[{ds_name}] {folder}"
+        file_list = self.file_list_in_folder(fcd_location.ds_ref(),
+                                             folder_path)
+        if delete_folder and file_list == []:
             self.delete_datastore_folder(ds_name, folder, dc_ref)
 
     def get_fcd_consumer(self, ds_ref, disk_id):
@@ -2568,6 +2571,46 @@ class VMwareVolumeOps(object):
 
     def _get_fcd_loc(self, prov_loc):
         return FcdLocation.from_provider_location(prov_loc)
+
+    def _get_browser(self, ds_ref):
+        value = ds_ref.value.replace('datastore', 'datastoreBrowser-datastore')
+        browser = vim_util.get_moref(value, 'HostDatastoreBrowser')
+        return browser
+
+    def _get_HostDatastoreBrowserSearchSpec(self, pattern, details=None):
+        cf = self._session.vim.client.factory
+        browser_spec = cf.create("ns0:HostDatastoreBrowserSearchSpec")
+        browser_spec.matchPattern = pattern
+        browser_spec.details = details
+        return browser_spec
+
+    def _get_subfolder_list(self, browser, folder_path, search_spec):
+        task = self._session.invoke_api(self._session.vim,
+                                        'SearchDatastoreSubFolders_Task',
+                                        browser, datastorePath=folder_path,
+                                        searchSpec=search_spec)
+        return task
+
+    def file_list_in_folder(self, ds_ref, folder_path):
+        def _get_task_detail(task_ref, session):
+            lst_props = ["info"]
+            props = session.invoke_api(vim_util, "get_object_properties_dict",
+                                       session.vim, task_ref, lst_props)
+            return props["info"]
+        browser = self._get_browser(ds_ref)
+        search_spec = self._get_HostDatastoreBrowserSearchSpec('*')
+        search_task = self._get_subfolder_list(browser, folder_path,
+                                               search_spec)
+        self._session.wait_for_task(search_task)
+        info = _get_task_detail(search_task, self._session)
+        res = []
+        if info.state == "success":
+            for i in info.result[0]:
+                if hasattr(i, 'file') and i.file != []:
+                    for j in i.file:
+                        res.append(i.folderPath + j.path)
+
+        return res
 
 
 class FcdLocation(object):
